+++ extension/src/setup/webview/messages.ts
@@ -14,19 +14,21 @@ import {
   RegisteredCliCommands,
   RegisteredCommands
 } from '../../commands/external'
-import { ConfigKey, setConfigValue } from '../../vscode/config'
 import { openUrl } from '../../vscode/external'
 
 export class WebviewMessages {
   private readonly getWebview: () => BaseWebview<TSetupData> | undefined
   private readonly initializeGit: () => void
+  private readonly updateStudioOffline: (offline: boolean) => Promise<void>
 
   constructor(
     getWebview: () => BaseWebview<TSetupData> | undefined,
-    initializeGit: () => void
+    initializeGit: () => void,
+    updateStudioOffline: (shareLive: boolean) => Promise<void>
   ) {
     this.getWebview = getWebview
     this.initializeGit = initializeGit
+    this.updateStudioOffline = updateStudioOffline
   }
 
   public sendWebviewMessage({
@@ -100,10 +102,7 @@ export class WebviewMessages {
           RegisteredCommands.REMOVE_STUDIO_ACCESS_TOKEN
         )
       case MessageFromWebviewType.SET_STUDIO_SHARE_EXPERIMENTS_LIVE:
-        return setConfigValue(
-          ConfigKey.STUDIO_SHARE_EXPERIMENTS_LIVE,
-          message.payload
-        )
+        return this.updateStudioOffline(message.payload)
       case MessageFromWebviewType.OPEN_EXPERIMENTS_WEBVIEW:
         return commands.executeCommand(RegisteredCommands.EXPERIMENT_SHOW)
       case MessageFromWebviewType.REMOTE_ADD: