+++ webview/src/setup/components/remotes/RemoteDetails.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import { RemoteList } from 'dvc/src/setup/webview/contract'
+import { SupportedStorage } from './SupportedStorage'
 import { MultiProjectRemotes } from './MultiProjectRemotes'
 import { ProjectRemotes } from './ProjectRemotes'
 import { EmptyState } from '../../../shared/components/emptyState/EmptyState'
@@ -33,6 +34,7 @@ export const RemoteDetails: React.FC<{
         onClick={removeRemote}
         text="Remove"
       />
+      <SupportedStorage />
     </EmptyState>
   )
 }